+++ nx-plugin-typebox/src/generators/executor/executor.ts
@@ -1,9 +1,10 @@
-import { formatFiles, joinPathFragments, Tree } from '@nx/devkit';
+import { formatFiles, joinPathFragments, names, Tree } from '@nx/devkit';
 import { ExecutorGeneratorSchema } from './schema';
 import { executorGenerator as nxExecutorGenerator } from '@nx/plugin/generators';
 import { relative } from 'path';
 import { ensureConfiguration } from '../configuration/configuration';
 import { findProjectByInnerPath } from '../utils/find-project-by-inner-path';
+import { determineArtifactNameAndDirectoryOptions } from '@nx/devkit/src/generators/artifact-name-and-directory-utils';
 
 export async function executorGenerator(
   tree: Tree,
@@ -12,14 +13,21 @@ export async function executorGenerator(
   const cwdRelativeToWorkspaceRoot = relative(tree.root, process.cwd());
   const executorRootPath = joinPathFragments(
     cwdRelativeToWorkspaceRoot,
-    options.path
+    options.path,
+    '..'
   );
   const projectName = findProjectByInnerPath(tree, executorRootPath);
   const config = await ensureConfiguration(tree, projectName);
 
+  const { artifactName: generatorName } =
+    await determineArtifactNameAndDirectoryOptions(tree, options);
+
+  const { className } = names(generatorName);
+
   const task = await nxExecutorGenerator(tree, {
     ...options,
     skipFormat: true,
+    unitTestRunner: options.unitTestRunner ?? 'jest',
   });
 
   tree.delete(joinPathFragments(executorRootPath, 'schema.d.ts'));
@@ -29,7 +37,9 @@ export async function executorGenerator(
     
     export const ${config.exportName} = Type.Object({
       name: Type.String(),
-    });`
+    });
+    
+    export type ${className}ExecutorSchema = typeof ${config.exportName};`
   );
 
   if (!options.skipFormat) {