+++ nx-plugin-typebox/src/schema-fragments/description.ts
@@ -2,8 +2,10 @@ import { Type } from '@sinclair/typebox';
 
 export const description = (identifier: string) =>
   Type.Object({
-    description: Type.String({
-      description: `${identifier} description.`,
-      'x-priority': 'important',
-    }),
+    description: Type.Optional(
+      Type.String({
+        description: `${identifier} description.`,
+        'x-priority': 'important',
+      })
+    ),
   });