+++ nx-plugin-typebox/src/schema-fragments/skip-format.ts
@@ -1,9 +1,11 @@
 import { Type } from '@sinclair/typebox';
 
 export const skipFormat = Type.Object({
-  skipFormat: Type.Boolean({
-    description: 'Skip formatting files.',
-    default: false,
-    'x-priority': 'internal',
-  }),
+  skipFormat: Type.Optional(
+    Type.Boolean({
+      description: 'Skip formatting files.',
+      default: false,
+      'x-priority': 'internal',
+    })
+  ),
 });