+++ nx-plugin-typebox/src/schema-fragments/skip-lint-checks.ts
@@ -1,8 +1,10 @@
 import { Type } from '@sinclair/typebox';
 
 export const skipLintChecks = Type.Object({
-  skipLintChecks: Type.Boolean({
-    default: false,
-    description: 'Do not add an eslint configuration for plugin json files.',
-  }),
+  skipLintChecks: Type.Optional(
+    Type.Boolean({
+      default: false,
+      description: 'Do not add an eslint configuration for plugin json files.',
+    })
+  ),
 });