+++ nx-plugin-typebox/src/schema-fragments/unit-test-runner.ts
@@ -1,9 +1,11 @@
 import { Type } from '@sinclair/typebox';
 
 export const unitTestRunner = Type.Object({
-  unitTestRunner: Type.Unsafe<'jest' | 'vitest' | 'none'>({
-    enum: ['jest', 'vitest', 'none'],
-    description: 'Test runner to use for unit tests.',
-    default: 'jest',
-  }),
+  unitTestRunner: Type.Optional(
+    Type.Unsafe<'jest' | 'vitest' | 'none'>({
+      enum: ['jest', 'vitest', 'none'],
+      description: 'Test runner to use for unit tests.',
+      default: 'jest',
+    })
+  ),
 });