+++ docs-site/docs/changelog.md
@@ -5,3 +5,12 @@ hide_title: true
 sidebar_position: 2
 slug: /changelog
 ---
+## 0.1.0 (2025-03-20)
+
+### 🚀 Features
+
+- **nx-plugin-typebox:** setup generators to wrap nx-plugin generators ([abd019a](https://github.com/AgentEnder/nx-plugin-typebox/commit/abd019a))
+
+### ❤️ Thank You
+
+- Craigory Coppola @AgentEnder